remove check for current comment
Because we are listening for update events and reloading the comments when create, updates and deletes occur we cannot check for the webix_comments_current class. This only was passing before because of an issue with the test object and new comments were not actually being saved to the database but only added temporarily to the UI. Therefore the listener for the datacollection did not fire and the ui did not rebuild itself with the "new" comment. So it looked good but was actually not working due to this issue and the tests were built around this faulty setup.